+++ tools/extraChart/tool.js
@@ -8,7 +8,7 @@
         name:'extraChart',
         category: 'analysis',
         type:'multiple',
-        possibleValues: ['material','principled','tension','potential','brilliant','smooth','gauge'],
+        possibleValues: ['material','principled','tension','potential','brilliant','smooth','gauge','local'],
         defaultValue: 'material,principled,tension,smooth,gauge,brilliant',
         advanced: true
       },
@@ -31,6 +31,7 @@
         'extraChart.principled': 'Principled',
         'extraChart.tension': 'Max tension',
         'extraChart.potential': 'Max potential',
+        'extraChart.local': 'Local eval',
         'extraChart.brilliant': 'Find interesting moves',
         'extraChart.smooth': 'Chart smoothing',
         'extraChart.gauge': 'on Eval gauge',
@@ -49,6 +50,7 @@
         'extraChart.principled': 'Principial',
         'extraChart.tension': 'Tensiune maxim\u0103',
         'extraChart.potential': 'Poten\u0163ial maxim',
+        'extraChart.local': 'Evaluare local\u0103',
         'extraChart.brilliant': 'G\u0103se\u015Fte mut\u0103ri interesante',
         'extraChart.smooth': 'Netezire grafice',
         'extraChart.gauge': 'pe bara de Eval',
@@ -65,6 +67,9 @@
       originalChart: '#D85000',
       materialChart: '#258F0B',
       principledChart: '#250B8F',
+      localChart: ()=>this.lichessTools.isDark()
+                              ? '#FFFF00'
+                              : '#705800',
       maxTensionLine: '#FF0000',
       maxPotentialLine: '#008000',
       interestingMoves: ()=>this.lichessTools.isDark()
@@ -514,6 +519,21 @@
         .filter(r=>!!r);
     };
 
+    getLocalData = (mainline) => {
+      const parent=this.lichessTools;
+      const Math=parent.global.Math;
+      return mainline
+        .map((node,x) => {
+          if (!node.ceval) return null;
+          const cp=node.ceval.cp||node.ceval.mate*9999||0;
+          return {
+            y: 2 / (1 + Math.exp(-0.004 * cp)) - 1,
+            x: x
+          };
+        })
+        .filter(r=>!!r);
+    };
+
     setBrilliant = (mainline,forced) => {
       if (!forced && mainline.brilliantInit) return;
       const parent=this.lichessTools;
@@ -634,7 +654,7 @@
       const $=parent.$;
       if (!this.options.brilliant) return;
 
-      const hcElem=$('div.computer-analysis.active #acpl-chart-container, div.study__server-eval.ready')[0];
+      const hcElem=$('#acpl-chart-container.lichessTools-extraChart, div.computer-analysis.active #acpl-chart-container, div.study__server-eval.ready')[0];
       let state=hcElem?.traverseState;
       if (!state || forced) {
         state=parent.traverse();
@@ -696,7 +716,7 @@
       const parent=this.lichessTools;
       const $=parent.$;
       const trans=parent.translator;
-      const container=$('div.computer-analysis.active #acpl-chart-container, div.study__server-eval.ready');
+      const container=$('#acpl-chart-container.lichessTools-extraChart, div.computer-analysis.active #acpl-chart-container, div.study__server-eval.ready');
       if (parent.inViewport(container)<=0) return;
       const chart=this._chart;
       if (!chart) return;
@@ -735,9 +755,10 @@
       const trans=parent.translator;
 
       if (!lichess.analysis) return;
-      if (!lichess.analysis.tree.root.eval&&!lichess.analysis.tree.root.children.at(0)?.eval) return;
-      const container=$('div.computer-analysis.active #acpl-chart-container, div.study__server-eval.ready');
-      if (parent.inViewport(container[0])<=0) return;
+      if (!lichess.analysis.mainline.find(n=>n.eval||n.ceval)) {
+         return;
+      }
+      let container=$('#acpl-chart-container.lichessTools-extraChart, div.computer-analysis.active #acpl-chart-container, div.study__server-eval.ready');
       let chart=this._chart;
       if (!chart) {
         chart = lichess.analysis.study?.serverEval?.chart;
@@ -748,11 +769,22 @@
             chart=await mod.acpl(canvas);
           }
         }
+        if (!chart&&this.options.local&&!lichess.analysis.study) {
+          const underboard=$('.analyse__underboard');
+          if (underboard.length) {
+            container=$('<div id="acpl-chart-container" class="lichessTools-extraChart"><canvas id="acpl-chart"></canvas></div>')
+              .appendTo(underboard);
+            const mod=await lichess.asset.loadEsm("chart.game");
+            const mainline=lichess.analysis.mainline;
+            chart = await mod.acpl($("#acpl-chart")[0], lichess.analysis.data, mainline, lichess.analysis.trans);
+          }
+        }
         if (chart && !this._chart) {
           this.setChart(chart);
         } 
       }
       if (!chart) return;
+      if (parent.inViewport(container[0])<=0) return;
       if (!this.options.needsChart) {
         $('div.lichessTools-chartInfo',container).remove();
       } else {
@@ -774,22 +806,32 @@
         existingMaterial=-1;
         updateChart=true;
       }
-      if (existingMaterial<0 && this.options.material) {
-        chart.data.datasets.push({
-          label:'Material',
-          type:'line',
-          data: this.smooth(this.getMaterialData(lichess.analysis.mainline)),
-          borderWidth: 2,
-          borderDash:[3,3],
-          cubicInterpolationMode: this.options.smooth?'monotone':undefined,
-          pointRadius: 0,
-          pointHitRadius: 0,
-          pointHoverRadius: 0,
-          borderColor: this.colors.materialChart,
-          order: 1,
-          datalabels: { display: false }
-        });
-        updateChart=true;
+      if (this.options.material) {
+        if (existingMaterial<0) {
+          chart.data.datasets.push({
+            label:'Material',
+            type:'line',
+            data: this.smooth(this.getMaterialData(lichess.analysis.mainline)),
+            borderWidth: 2,
+            borderDash:[3,3],
+            cubicInterpolationMode: this.options.smooth?'monotone':'default',
+            tension: 0,
+            pointRadius: 0,
+            pointHitRadius: 0,
+            pointHoverRadius: 0,
+            borderColor: this.colors.materialChart,
+            order: 1,
+            datalabels: { display: false }
+          });
+          updateChart=true;
+        } else {
+          const dataset=chart.data.datasets[existingMaterial];
+          if (dataset.data?.length<lichess.analysis.mainline.length) {
+            const newData = this.smooth(this.getMaterialData(lichess.analysis.mainline));
+            dataset.data=newData;
+            updateChart=true;
+          }
+        }
       }
 
       let existingPrincipled = chart.data.datasets.findIndex(s=>s.label==='Principled');
@@ -798,22 +840,75 @@
         existingPrincipled=-1;
         updateChart=true;
       }
-      if (existingPrincipled<0 && this.options.principled) {
-        chart.data.datasets.push({
-          label:'Principled',
-          type:'line',
-          data: this.smooth(this.getPrincipledData(lichess.analysis.mainline)),
-          borderWidth: 2,
-          borderDash:[3,3],
-          cubicInterpolationMode: this.options.smooth?'monotone':undefined,
-          pointRadius: 0,
-          pointHitRadius: 0,
-          pointHoverRadius: 0,
-          borderColor: this.colors.principledChart,
-          order: 1,
-          datalabels: { display: false }
-        });
+      if (this.options.principled) {
+        if (existingPrincipled<0) {
+          chart.data.datasets.push({
+            label:'Principled',
+            type:'line',
+            data: this.smooth(this.getPrincipledData(lichess.analysis.mainline)),
+            borderWidth: 2,
+            borderDash:[3,3],
+            cubicInterpolationMode: this.options.smooth?'monotone':'default',
+            tension: 0,
+            pointRadius: 0,
+            pointHitRadius: 0,
+            pointHoverRadius: 0,
+            borderColor: this.colors.principledChart,
+            order: 1,
+            datalabels: { display: false }
+          });
+          updateChart=true;
+        } else {
+          const dataset=chart.data.datasets[existingPrincipled];
+          if (dataset.data?.length<lichess.analysis.mainline.length) {
+            const newData = this.smooth(this.getPrincipledData(lichess.analysis.mainline));
+            dataset.data=newData;
+            updateChart=true;
+          }
+        }
+      }
+
+      let existingLocal = chart.data.datasets.findIndex(s=>s.label==='Local');
+      if (existingLocal>=0 && (this.prevSmooth!=this.options.smooth || !this.options.local)) {
+        chart.data.datasets.splice(existingLocal,1);
+        existingLocal=-1;
         updateChart=true;
+      }
+      if (this.options.local) {
+        const mainline=lichess.analysis.tree.getNodeList(lichess.analysis.node.path||'');
+        let lastNode=mainline.at(-1)?.children[0];
+        while (lastNode) {
+          mainline.push(lastNode);
+          lastNode=lastNode.children[0];
+        }
+        if (existingLocal<0) {
+          chart.data.datasets.push({
+            label:'Local',
+            type:'line',
+            data: this.smooth(this.getLocalData(mainline)),
+            borderWidth: 2,
+            borderDash:[1,5],
+            cubicInterpolationMode: this.options.smooth?'monotone':'default',
+            tension: 0,
+            pointRadius: 0,
+            pointHitRadius: 0,
+            pointHoverRadius: 0,
+            borderColor: this.colors.localChart,
+            fill: {
+              target: 'start',
+              above: this.colors.localChart()+'20'
+            },
+            order: 1,
+            datalabels: { display: false }
+          });
+          updateChart=true;
+        } else {
+          const dataset=chart.data.datasets[existingLocal];
+          const existingData = dataset.data;
+          const newData = this.smooth(this.getLocalData(mainline));
+          updateChart=JSON.stringify(existingData)!=JSON.stringify(newData);
+          if (updateChart) dataset.data=newData;
+        }
       }
 
       let existingMaxTension = chart.data.datasets.findIndex(s=>s.label==='Max tension');
@@ -822,7 +917,8 @@
         existingMaxTension=-1;
         updateChart=true;
       }
-      if (existingMaxTension<0 && this.options.tension) {
+      if (this.options.tension) {
+        if (existingMaxTension<0) {
         const x=this.getMaxTension(lichess.analysis.mainline);
         chart.data.datasets.push({
           label:'Max tension',
@@ -846,6 +942,18 @@
           }
         });
         updateChart=true;
+        } else {
+          const dataset=chart.data.datasets[existingMaxTension];
+          const existingData = dataset.data;
+          if (this.prevMainlineLength!=lichess.analysis.mainline.length) {
+            const x=this.getMaxTension(lichess.analysis.mainline);
+            dataset.data=[
+              { x: x, y:-1.05 },
+              { x: x, y:1.05 }
+            ];
+            updateChart=true;
+          }
+        }
       }
 
       let existingMaxPotential = chart.data.datasets.findIndex(s=>s.label==='Max potential');
@@ -854,7 +962,8 @@
         existingMaxPotential=-1;
         updateChart=true;
       }
-      if (existingMaxPotential<0 && this.options.potential) {
+      if (this.options.potential) {
+        if (existingMaxPotential<0) {
         const x=this.getMaxPotential(lichess.analysis.mainline);
         chart.data.datasets.push({
           label:'Max potential',
@@ -878,9 +987,20 @@
           }
         });
         updateChart=true;
+        } else {
+          const dataset=chart.data.datasets[existingMaxPotential];
+          if (this.prevMainlineLength!=lichess.analysis.mainline.length) {
+            const x=this.getMaxPotential(lichess.analysis.mainline);
+            dataset.data=[
+              { x: x, y:-1.05 },
+              { x: x, y:1.05 }
+            ];
+            updateChart=true;
+          }
+        }
       }
 
-      if (this.options.brilliant) {
+      if (this.options.brilliant && !$('#acpl-chart-container').is('.lichessTools-extraChart')) {
         this.setBrilliant(lichess.analysis.mainline,forced);
 
         this.showGoodMoves(forced);
@@ -890,7 +1010,11 @@
         await this.showChristmasTree();
       }
 
-      if (updateChart) chart.update('none');
+      if (updateChart) {
+        chart.options.scales.x.max=Math.max.apply(null,chart.data.datasets.map(ds=>ds.data.map(p=>p.x)).flat());
+        chart.update('none');
+        this.prevMainlineLength=lichess.analysis.mainline.length;
+      }
       this.prevSmooth=this.options.smooth;
     };
 
@@ -966,8 +1090,9 @@
         tension:parent.isOptionSet(value,'tension'),
         potential:parent.isOptionSet(value,'potential'),
         brilliant:parent.isOptionSet(value,'brilliant'),
+        local:parent.isOptionSet(value,'local'),
         smooth:parent.isOptionSet(value,'smooth'),
-        get needsChart() { return this.material || this.principled || this.tension || this.brilliant; },
+        get needsChart() { return this.material || this.principled || this.tension || this.brilliant || this.local; },
         gauge:parent.isOptionSet(value,'gauge'),
         christmas:!!parent.currentOptions.getValue('christmas')
       };
@@ -980,6 +1105,13 @@
       if (!this.options.gauge) {
         $('div.eval-gauge tick.lichessTools-material,div.eval-gauge tick.lichessTools-principled').remove();
         this.prevFen=null;
+      }
+      if (!this.options.local) {
+        const container=$('#acpl-chart-container.lichessTools-extraChart');
+        if (container.length) {
+          container.remove();
+          this._chart=null;
+        }
       }
       this.interval=parent.global.setInterval(()=>{
         this.generateCharts();